test(vitest): move no-worker selection coverage to integration
